--- src/features/favorites/components/AdminFavorites/adminFavorites.styles.ts
@@ -1,9 +1,0 @@
-import { Box, styled } from "@mui/material";
-
-export const Container = styled(Box)(() => ({
-  padding: '16px',
-  display: 'flex',
-  flexDirection: 'column',
-  alignItems: 'center',
-  width: '100%'
-}));
